+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+## 1.2.0 - 2021-05-27
+### Added
+- Optimizers: Final graph now contains invalid points as well (if present).
+- Selectors: When using negative selection, leave the number of points on the selector.
+- Selectors: Curvature selector by Ondra Benedikt.
+- Script 'ng_generate_data' for generating numpy files for the algorithm from an image of the map.
+
+### Changed
+- Optimizers: Individual plotting is now forced on the main figure, instead of the global one.
 
 ## 1.1.1 - 2021-05-20
 ### Added